rename error display
