+++ src/app/app.module.js
@@ -10,7 +10,7 @@ import UiRouterConfig from './configuration/ui-router.config';
 import MainComponent from './app.component';
 import Header from './header/header.module';
 import Footer from './footer/footer.module';
-import ErrorDisplay from './error-display/error-display.module';
+import ErrorDisplay from './error/error.module';
 import CreateQuestion from './create-question/create-question.module';
 import QuestionTrack from './question-track/question-track.module';
