+++ src/app/error/error.module.js
@@ -1,16 +1,16 @@
 import angular from 'angular';
-import errorDisplayComponent from './error-display.component';
-import './error-display.css';
+import errorComponent from './error.component';
+import './error.css';
 import './error.jpg';
 
 
 // Module which has the error display component.  
 
-const moduleName = 'unlock-the-truth.error-display';
+const moduleName = 'unlock-the-truth.error';
 
 angular
     .module(moduleName, [])
-    .component('uttErrorDisplay', errorDisplayComponent)
+    .component('uttError', errorComponent)
 
     // Configure the routing for this module. 
     .config($stateProvider => {
@@ -19,7 +19,7 @@ angular
         // Register the home state, and add the home component to it. 
         $stateProvider.state('error', {
             url: '/error',
-            component: 'uttErrorDisplay',
+            component: 'uttError',
             params: {
                 // allow this state to be transitioned to with an error param. 
                 error: null
