+++ src/app/question-track/question-track.controller.js
@@ -10,11 +10,12 @@ import {
 // Controller for the question track component.  
 
 class QuestionTrackController {
-    constructor($scope, uttQuestionTrackService) {
+    constructor($scope, $state, uttQuestionTrackService) {
         'ngInject';
 
         // Members
         this._$scope = $scope;
+        this._$state = $state;
         this._questionTrackService = uttQuestionTrackService;
         this._answerResult = null;
 
@@ -43,7 +44,9 @@ class QuestionTrackController {
                 this.model = result;
             })
             .catch(error => {
-                this.error = error;
+                this._$state.go('error', {
+                    error: error
+                });
             })
             .finally(() => {
                 this.loading = false;
@@ -96,12 +99,20 @@ class QuestionTrackController {
 
     // Check whether the user has unlocked all of the questions. 
     allQuestionsUnlocked() {
-        if(!this.model){
+        if (!this.model) {
             return false;
         }
 
         // No current question? No locked questions? Seems like we got em all. 
         return !this.model.currentQuestion && this.model.lockedQuestions.length === 0;
+    }
+
+    testErrorState() {
+        this._$state.go('error', {
+            error: {
+                message: 'BAD BAD BAD'
+            }
+        });
     }
 }
