+++ src/app/question-track/question-track.html
@@ -30,3 +30,9 @@
         </div>
     </div>
 </div>
+
+<div class="row">
+    <div class="col-xs-12">
+        <button class="btn btn-primary" ng-click="$ctrl.testErrorState()">ERROR</button>
+    </div>
+</div>
